requirements.txt
python-dotenv==1.0.0
openai==1.12.0
supabase==2.3.4
pyyaml==6.0.1

